Change models to follow mediator pattern
